Moved playerRelated things to PlayerMaster from GameMaster
